feat: [backend] Creates sequelize Seeders

+++ backend/src/database/migrations/20221117004019-create-accounts.js
@@ -10,7 +10,7 @@ module.exports = {
         autoIncrement: true,
       },
       balance: {
-        type: Sequelize.INTEGER,
+        type: Sequelize.DECIMAL(12, 2),
         allowNull: false,
       },
     })
